feat(kyc): say what the hosted check needs before handing over
The bridge-hosted task (Bridge's hosted flow, Persona) was the one
verification path that handed the user to the vendor with no prep at
all: tap "Complete verification" on the pending-tasks card and the
browser sheet opens on step one. It is also the path where that hurts
most — unlike the Sumsub SDK, the hosted flow keeps no partial
progress, so anyone who leaves mid-check to find a proof of address
loses everything and restarts.

Put a screen in front of it. It lists the documents to have in hand,
warns that the check has to be finished in one go, and says how long
it takes; the CTA owns the handoff. A page rather than a sheet because
the prep is the whole content — the modal it started as needed
scrolling to reach its own CTA, and centred in the overlay it
overflowed past the top of the scroll container where nothing could
scroll it back.

The launch itself moves out of the card into useBridgeHostedVerification
unchanged: the tab is still reserved synchronously inside the click
(a window.open after the ~800ms start-action round-trip is popup-blocked
on Safari), opener is still severed, and the return still refetches off
browserFinished on native / visibilitychange on web.

Also carries two fixes into the shared checklist: the how-long note
loses its frame (it read as one more requirement rather than a note),
and the bordered list takes its own surface — it was only ever white
because the modal panel behind it was.
