+++ src/app/(mobile-ui)/profile/identity-verification/additional/page.tsx
@@ -1,0 +1,5 @@
+import { AdditionalVerificationView } from '@/components/Kyc/AdditionalVerificationView'
+
+export default function AdditionalVerificationPage() {
+    return <AdditionalVerificationView />
+}
