+++ src/components/Home/PendingVerificationTasks.tsx
@@ -2,7 +2,7 @@
 
 import { useCallback, useEffect, useState } from 'react'
 import { useTranslations } from 'next-intl'
-import { startBridgeHostedVerification } from '@/app/actions/sumsub'
+import { useRouter } from 'next/navigation'
 import { Button } from '@/components/0_Bruddle/Button'
 import Carousel from '@/components/Global/Carousel'
 import { Icon } from '@/components/Global/Icons/Icon'
@@ -11,7 +11,6 @@ import { useAuth } from '@/context/authContext'
 import { useCapabilities } from '@/hooks/useCapabilities'
 import type { NextAction } from '@/types/capabilities'
 import { bridgeTaskDismissalKey, selectBridgeTasks } from '@/utils/bridge-tasks.utils'
-import { isNativeBridge, openExternalUrl } from '@/utils/capacitor'
 import { formatEffectiveDate } from '@/utils/format.utils'
 import { getUserPreferences, updateUserPreferences } from '@/utils/general.utils'
 import Card from '../Global/Card'
@@ -58,7 +57,8 @@ function taskCopy(task: NextAction): { title: string; description: string } {
  * The ToS flow is SNAPSHOTTED at tap time: the task list re-derives from every
  * user refetch (~4s auto-refresh while rails are pending), and the open modal
  * must survive its task disappearing mid-flow — the card hides, the flow keeps
- * running. The hosted flow leaves the app entirely (see handleOpenTask).
+ * running. The hosted flow hands off to the vendor, so it goes through the
+ * additional-verification screen first (see AdditionalVerificationView).
  *
  * `dismissible` (the /home mount): each ADVISORY (future-dated) slide carries
  * its own X that dismisses ONLY that task — the other slides stay. Dismissals
@@ -72,11 +72,9 @@ function taskCopy(task: NextAction): { title: string; description: string } {
 export default function PendingVerificationTasks({ dismissible = false }: { dismissible?: boolean }) {
     const t = useTranslations('home')
     const { nextActions } = useCapabilities()
-    const { user, fetchUser } = useAuth()
+    const { user } = useAuth()
     const [activeTosTask, setActiveTosTask] = useState<NextAction | null>(null)
-    const [isStartingHosted, setIsStartingHosted] = useState(false)
-    const [awaitingReturn, setAwaitingReturn] = useState(false)
-    const [error, setError] = useState<string | null>(null)
+    const router = useRouter()
     // Stored dismissals, tagged with the user they were loaded for
     // (localStorage is unreadable during SSR, hence the post-render effect).
     // The dismissible mount must not paint until the CURRENT user's entry is
@@ -87,17 +85,6 @@ export default function PendingVerificationTasks({ dismissible = false }: { dism
 
     const userId = user?.user?.userId
     const tasks = selectBridgeTasks(nextActions)
-    const hasHostedTask = tasks.some((task) => task.kind === 'bridge-hosted')
-    const taskKeys = tasks
-        .map((task) => task.key)
-        .sort()
-        .join(',')
-
-    // A new task set means the previous failure context is gone.
-    useEffect(() => {
-        setError(null)
-    }, [taskKeys])
-
     useEffect(() => {
         if (!dismissible || !userId) return
         // Pre-fingerprint native builds (≤1.0.50) persisted this preference as a
@@ -145,135 +132,19 @@ export default function PendingVerificationTasks({ dismissible = false }: { dism
           )
 
     const handleOpenTask = useCallback(
-        async (task: NextAction) => {
-            setError(null)
+        (task: NextAction) => {
             if (task.kind === 'accept-tos') {
                 setActiveTosTask(task)
                 return
             }
-            // NOT an iframe: `bridge.withpersona.com` serves
-            // `X-Frame-Options: SAMEORIGIN`, so embedding it rendered
-            // "refused to connect" for EVERY user. It has to be a real
-            // top-level page (native: the in-app browser). Bridge's ToS link
-            // (`compliance.bridge.xyz`) sends no framing header, which is why
-            // BridgeTosStep keeps its iframe.
-            //
-            // On web, reserve the tab HERE — synchronously, inside the click's
-            // user-activation window. Fetching the link takes ~800ms, and a
-            // window.open() after that await is no longer gesture-initiated,
-            // so Safari/Firefox block it — the same "nothing happens" symptom
-            // this PR exists to fix. The reserved tab is navigated once the
-            // URL lands, and closed if it never does.
-            // isNativeBridge(), not isCapacitor(): the latter is also true for
-            // any build carrying NEXT_PUBLIC_CAPACITOR_BUILD (vercel previews),
-            // where the native apis don't exist and we'd skip the reservation
-            // in a real browser.
-            const native = isNativeBridge()
-            const reservedTab = native ? null : window.open('', '_blank')
-            // The reserved tab can't carry `noopener` (that returns null and
-            // defeats the reservation), so sever the back-reference by hand —
-            // otherwise Persona, and anything it redirects to, holds a handle
-            // that can navigate the signed-in tab (reverse tabnabbing).
-            if (reservedTab) reservedTab.opener = null
-
-            setIsStartingHosted(true)
-            let url: string | undefined
-            try {
-                ;({ url } = await startBridgeHostedVerification())
-            } catch (error) {
-                // The action body catches its own errors, but a server action
-                // can still REJECT at the transport layer — a dropped network,
-                // or a deploy invalidating the action id mid-flight. Without
-                // this the button stays on "Loading..." forever and the blank
-                // reserved tab is orphaned.
-                reservedTab?.close()
-                console.error('[pending-tasks] start-action rejected', error)
-                setError("We couldn't start the verification. Please try again in a moment.")
-                return
-            } finally {
-                setIsStartingHosted(false)
-            }
-            if (!url) {
-                // Friendly copy regardless of the server detail (a 403 here just
-                // means the action aged out); refetch so a stale card self-corrects.
-                reservedTab?.close()
-                setError("We couldn't start the verification. Please try again in a moment.")
-                void fetchUser().catch(() => undefined)
-                return
-            }
-            try {
-                if (native) {
-                    await openExternalUrl(url)
-                } else if (reservedTab && !reservedTab.closed) {
-                    // `.closed` matters: assigning href to a closed window is a
-                    // silent no-op, so without this the user would tap and see
-                    // nothing — the very failure this PR removes.
-                    reservedTab.location.href = url
-                } else {
-                    // No usable tab: pop-ups blocked, a standalone PWA, or the
-                    // user closed the blank tab while we fetched. Same-tab
-                    // navigation is never gesture-gated, so it always lands.
-                    reservedTab?.close()
-                    window.location.href = url
-                    return
-                }
-            } catch (error) {
-                reservedTab?.close()
-                console.error('[pending-tasks] failed to open Bridge hosted verification', error)
-                setError("We couldn't open the verification. Please try again in a moment.")
-                return
-            }
-            setAwaitingReturn(true)
+            // The hosted flow gets its own screen first. It runs at the vendor,
+            // in a browser we don't control, and keeps no partial progress — a
+            // user who leaves mid-check to find a document restarts from step
+            // one. That is a page's worth of prep, and it owns the handoff.
+            router.push('/profile/identity-verification/additional')
         },
-        [fetchUser]
+        [router]
     )
-
-    // Nothing polls for this cohort — the ~4s user auto-refresh only runs
-    // while a rail is `pending`, and these are `requires-info` — so pick the
-    // result up when the user comes back from the hosted flow.
-    //
-    // Deliberately NOT one-shot: the first return is often incidental (a quick
-    // tab switch back, or Persona telling them to continue on their phone).
-    // Burning the single refetch there would leave the card up forever, so we
-    // keep listening and stop only once the task is actually gone (below).
-    useEffect(() => {
-        if (!awaitingReturn) return
-        const refresh = () => void fetchUser().catch(() => undefined)
-
-        if (isNativeBridge()) {
-            // Android WebViews don't reliably fire `visibilitychange` on
-            // resume — the same defect that makes useNativePlugins drive
-            // TanStack's focusManager off `appStateChange`. The in-app
-            // browser's own close event is the precise signal here.
-            let disposed = false
-            let remove: (() => void) | undefined
-            void import('@capacitor/browser')
-                .then(({ Browser }) => Browser.addListener('browserFinished', refresh))
-                .then((handle) => {
-                    // Cleanup can run while the dynamic import is still in
-                    // flight; without this the listener registers after the
-                    // fact and nobody ever removes it.
-                    if (disposed) handle.remove()
-                    else remove = () => handle.remove()
-                })
-                .catch((error) => console.error('[pending-tasks] browserFinished listener failed', error))
-            return () => {
-                disposed = true
-                remove?.()
-            }
-        }
-
-        const onReturn = () => {
-            if (document.visibilityState === 'visible') refresh()
-        }
-        document.addEventListener('visibilitychange', onReturn)
-        return () => document.removeEventListener('visibilitychange', onReturn)
-    }, [awaitingReturn, fetchUser])
-
-    // Stop listening once the hosted task clears — that's the success signal.
-    useEffect(() => {
-        if (awaitingReturn && !hasHostedTask) setAwaitingReturn(false)
-    }, [awaitingReturn, hasHostedTask])
 
     const closeTos = useCallback(() => setActiveTosTask(null), [])
 
@@ -319,21 +190,15 @@ export default function PendingVerificationTasks({ dismissible = false }: { dism
                                             variant="purple"
                                             shadowSize="4"
                                             className="mt-1 w-full"
-                                            disabled={isStartingHosted}
                                             onClick={() => handleOpenTask(task)}
                                         >
-                                            {isHosted
-                                                ? isStartingHosted
-                                                    ? 'Loading...'
-                                                    : 'Complete verification'
-                                                : 'Review terms'}
+                                            {isHosted ? 'Complete verification' : 'Review terms'}
                                         </Button>
                                     </div>
                                 </Card>
                             )
                         })}
                     </Carousel>
-                    {error && <p className="mt-2 text-center text-body-s text-error">{error}</p>}
                 </div>
             )}
