+++ src/i18n/app/messages/pt-BR.json
@@ -2855,15 +2855,28 @@
                 "questions": {
                     "title": "Algumas perguntas rápidas",
                     "body": "Declarações regulatórias que somos obrigados a pedir."
+                },
+                "proofOfAddress": {
+                    "title": "Um comprovante de residência, se você tiver",
+                    "body": "Uma conta de consumo, um extrato bancário ou uma carta oficial dos últimos 3 meses com seu nome e seu endereço. Nem todo mundo precisa enviar, e depois não dá para adicionar."
                 }
             },
             "howLongLabel": "Quanto tempo leva",
             "howLong": {
                 "standard": "Menos de 2 minutos para a maioria das pessoas. Se um revisor precisar analisar, de 1 a 3 dias úteis.",
-                "extended": "Mais do que a verificação comum. Se um revisor precisar analisar, de 1 a 3 dias úteis."
+                "extended": "Mais do que a verificação comum. Se um revisor precisar analisar, de 1 a 3 dias úteis.",
+                "hosted": "Cerca de 5 minutos com seus documentos em mãos. Se um revisor precisar analisar, de 1 a 3 dias úteis."
             },
             "extraDocNote": "Algumas pessoas precisam enviar mais um documento, como um comprovante de residência. Se for o seu caso, avisaremos antes de abrir a câmera.",
-            "startCta": "Tenho tudo, começar"
+            "startCta": "Tenho tudo, começar",
+            "singleSession": {
+                "title": "Faça tudo de uma vez só",
+                "body": "Nosso parceiro não salva seu progresso. Se você fechar a página ou sair para procurar um documento, começa de novo da primeira etapa."
+            }
+        },
+        "hostedPrep": {
+            "title": "O que esperar",
+            "description": "Nosso parceiro de pagamentos precisa de mais algumas coisas antes que suas transferências bancárias funcionem."
         }
     },
     "identity": {
